Add cat support for OKR database
Adds support for an optional OKR database provided as a CSV-file.

Initially also adds support in `cat` for an optional `--okr-db`
parameter. When provided, titles, objectives and projects will be
overwritten in the report with data from the database. If no KR ID is
available the database will be used to look it up based on the title.