Update install instructions to pin main branch
The release branch hasn't been updated in a while and recent changes to
okra are needed to parse weekly reports with WIs in them. This changes
the install instructions to pin the main branch so users get access to
these features when installing okra according to the instructions in the
readme.